deploy_package/requirements.txt
# 基础依赖
numpy>=1.20.0
fastapi==0.68.0
uvicorn==0.15.0
pydantic>=1.8.2

# 安全认证
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.5

# 可视化扩展
streamlit==1.28.0
plotly==5.18.0
dash==2.14.0
dash-bootstrap-components==1.4.0

# 实时数据
websockets==10.4
redis==4.5.5

# 生产环境
gunicorn==20.1.0
uvloop==0.17.0